Fix handling of root/merge commits with PGP signatures
Previously, the PGP signature detection assumed that a signature, if present, always appears on the fifth line of the commit object. However, this only works if a commit has exactly one parent, which isn't necessarily the case (a root commit has zero parents, and a merge commit has multiple parents). The parsing now determines whether the signature is on a line that starts with "gpgsig " and appears before the commit message starts.