Stop precompiling OpenCL kernel
Sadly, OpenCL kernel binaries don't have any ABI compatibility guarantees, which was causing the precompiled kernel in the binary to break after OS upgrades. So we're back to compiling the kernel at runtime.

There are IR formats for OpenCL kernels, but there doesn't seem to be a good compatibility story for them (SPIR is used in most places, but Apple doesn't support it and recommends LLVM bytecode instead).